+++ subcommand/controller/command_ent_test.go
@@ -38,3 +38,15 @@ func TestUpsertAnonymousTokenPolicyEnt(t *testing.T) {
 		},
 	})
 }
+
+func TestUpsertMeshGatewayTokenPolicyAndRole(t *testing.T) {
+	testUpsertMeshGatewayPolicyAndRole(t, map[string]gatewayTokenTest{
+		"test creation": {
+			partitionsEnabled: true,
+		},
+		"test creation with non default partition": {
+			partitionsEnabled:      true,
+			useNonDefaultPartition: true,
+		},
+	})
+}